pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "joblib>=1.5.2",
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "pandas>=2.3.2",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.1",
    "seaborn>=0.13.2",
]
